PUT Person fix

+++ src/main/java/home/holymiko/InvestmentScraperApp/Server/Type/DTO/create/PersonAccountCreateDTO.java
@@ -9,7 +9,7 @@ public class PersonAccountCreateDTO extends PersonDTO {
     // Credentials & Role
     private final AccountCreateDTO account;
 
-    public PersonAccountCreateDTO(Long id, String firstName, String middleName, String lastName, String email, int phone, AccountCreateDTO account) {
+    public PersonAccountCreateDTO(Long id, String firstName, String middleName, String lastName, String email, Long phone, AccountCreateDTO account) {
         super(id, firstName, middleName, lastName, email, phone);
         this.account = account;
     }
